fadc-optimization/requirements.txt
Flask==2.2.5
Pillow
